+++ src/BT/CustomLuaNodeState.hpp
@@ -2,6 +2,9 @@
 #ifndef BT_CUSTOM_LUA_NODE_STATE
 #define BT_CUSTOM_LUA_NODE_STATE
 
+#define BT_CUSTOM_LUA_NODE_STATE_REGISTRY_INDEX \
+    "BehaviorTree_CustomLuaNodeState"
+
 #include <memory>
 #include <lua.hpp>
 
@@ -16,13 +19,16 @@ struct CustomLuaNodeState
     CustomLuaNodeState();
 
     CustomLuaNode* cln;
-    void* vptr;
 };
 
 int getChildrenSize(lua_State* L);
 int activateChild(lua_State* L);
 
-int writerForTransferringFunction(lua_State* /*L*/, const void* p, std::size_t sz, void* ud);
+int writerForTransferringFunction(
+    lua_State* /*L*/,
+    const void* p,
+    std::size_t sz,
+    void* ud);
 
 } // namespace BT